refactorización y optimización del código

+++ migrations/20260207223632-create-precio.js
@@ -26,7 +26,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       precio: {
-        type: Sequelize.DOUBLE
+        type: Sequelize.DECIMAL(10, 2)
       }
     });
   },
